backend/requirements.txt
aniso8601
attrs
blinker
certifi
chardet
charset-normalizer
click
Flask
Flask-Cors
Flask-Migrate
flask-restx
Flask-Script
Flask-SQLAlchemy
greenlet
idna
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
Mako
MarkupSafe
Pillow
PyJWT
pyrsistent
python-dateutil
python-editor
pytz
referencing
requests
rpds-py
six
SQLAlchemy
typing_extensions
urllib3
uWSGI
Werkzeug
